feat: move generated proto types into pkg/wgpb
