+++ pkg/node/node.go
@@ -15,25 +15,24 @@ import (
 	"strings"
 	"time"
 
+	"github.com/gorilla/mux"
 	"github.com/jensneuse/abstractlogger"
+	"github.com/libp2p/go-reuseport"
 	"github.com/pires/go-proxyproto"
 	"github.com/valyala/fasthttp"
+	"golang.org/x/crypto/acme"
+	"golang.org/x/crypto/acme/autocert"
+	"golang.org/x/net/idna"
+	"golang.org/x/time/rate"
+
 	"github.com/wundergraph/wundergraph/pkg/apihandler"
 	"github.com/wundergraph/wundergraph/pkg/engineconfigloader"
 	"github.com/wundergraph/wundergraph/pkg/hooks"
 	"github.com/wundergraph/wundergraph/pkg/loadvariable"
 	"github.com/wundergraph/wundergraph/pkg/pool"
 	"github.com/wundergraph/wundergraph/pkg/validate"
+	"github.com/wundergraph/wundergraph/pkg/wgpb"
 	"github.com/wundergraph/wundergraph/pkg/wundernodeconfig"
-	"github.com/wundergraph/wundergraph/types/go/wgpb"
-	"golang.org/x/crypto/acme"
-	"golang.org/x/crypto/acme/autocert"
-	"golang.org/x/net/idna"
-	"golang.org/x/time/rate"
-
-	"github.com/libp2p/go-reuseport"
-
-	"github.com/gorilla/mux"
 )
 
 func New(ctx context.Context, info BuildInfo, cfg *wundernodeconfig.Config, log abstractlogger.Logger) *Node {
@@ -415,7 +414,7 @@ func (n *Node) startServer(nodeConfig wgpb.WunderNodeConfig) error {
 	if n.cfg.Server.ListenTLS {
 		manager := autocert.Manager{
 			Prompt: autocert.AcceptTOS,
-			//Cache:  certCache,
+			// Cache:  certCache,
 			HostPolicy: func(ctx context.Context, host string) error {
 				for _, allowedHost := range allowedHosts {
 					if allowedHost == host {
