+++ pkg/postresolvetransform/postresolvetransform.go
@@ -4,7 +4,8 @@ import (
 	"fmt"
 
 	"github.com/buger/jsonparser"
-	"github.com/wundergraph/wundergraph/types/go/wgpb"
+
+	"github.com/wundergraph/wundergraph/pkg/wgpb"
 )
 
 type Transformer struct {
