+++ pkg/wundergraphapi/client.go
@@ -11,7 +11,8 @@ import (
 	"github.com/andybalholm/brotli"
 	"github.com/buger/jsonparser"
 	"github.com/jensneuse/abstractlogger"
-	"github.com/wundergraph/wundergraph/types/go/wgpb"
+
+	"github.com/wundergraph/wundergraph/pkg/wgpb"
 )
 
 type Client struct {
