+++ types/generate.sh
@@ -1,11 +1,9 @@
 #!/usr/bin/env sh
 
-OUT_DIR="./go/wgpb"
-
-mkdir -p ${OUT_DIR}
-
 protoc  \
   --proto_path=./protos \
-  --go_out=. \
+  --go_out=.. \
+  --go_opt=module=github.com/wundergraph/wundergraph \
+  --go_opt=paths=import \
   --experimental_allow_proto3_optional \
   wundernode_config.proto
