--- types/go.mod
@@ -1,5 +1,0 @@
-module github.com/wundergraph/wundergraph/types
-
-go 1.18
-
-require google.golang.org/protobuf v1.28.0 // indirect
